Pocket-Rag/backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
qdrant-client
langchain-qdrant
pydantic
python-dotenv
pypdf